refactor: remove deprecated username server action
lib/server/username.ts was a simplified duplicate of lib/server/loginname.ts,
marked for deletion but still imported by lib/oidc.ts and lib/saml.ts for
session re-authentication during expired OIDC/SAML flows.

Both call sites now import sendLoginname from lib/server/loginname.ts, which
handles all the same cases plus IDP routing, passkey handling, and org discovery.
